Removed redundant code
